Add special handling to the STOP statement.
QB64, unlike QB4.5, treats the STOP statement the same way it treats the SYSTEM statement; as QB64 code isn't interpreted, but compiled, there's no way to go back to the IDE and make changes, hence the current behavior for the STOP statement. With this update, vWATCH64 treats STOP as it was originally intended, turning it into a permanent breakpoint, hardcoded into your program. You can't toggle this special breakpoint off, but you can use the "skip this line" feature to temporarily set it off.